CLOUD-747 Organized exception classes in the service layer
